+++ Modules/Intent.Modules.AspNetCore.Controllers/Templates/ExceptionFilter/ExceptionFilterTemplatePartial.cs
@@ -119,7 +119,8 @@ public partial class ExceptionFilterTemplate : CSharpTemplateBase<object>, ICSha
             .FindStatement(p => p.HasMetadata("exception-switch"));
     }
 
-    [IntentManaged(Mode.Fully)] public CSharpFile CSharpFile { get; }
+    [IntentManaged(Mode.Fully)]
+    public CSharpFile CSharpFile { get; }
 
     [IntentManaged(Mode.Fully)]
     protected override CSharpFileConfig DefineFileConfig()